Split files up, add prelude
