+++ event-store/src/event_context.rs
@@ -1,0 +1,18 @@
+use chrono::prelude::*;
+use serde_json::Value as JsonValue;
+
+/// Event context
+///
+/// Contains metadata for event and, most importantly, the creation time
+// TODO: Make `pub` -> `crate` when it's stabilised
+#[derive(Serialize, Deserialize)]
+pub struct EventContext {
+    /// TODO: What is this?
+    pub action: Option<String>,
+
+    /// Optional event "subject" or metadata
+    pub subject: Option<JsonValue>,
+
+    /// Event creation time
+    pub time: DateTime<Utc>,
+}
