+++ event-store/src/lib.rs
@@ -22,269 +22,24 @@ extern crate r2d2_postgres;
 extern crate tokio;
 
 pub mod adapters;
+mod aggregator;
+mod event;
+mod event_context;
+pub mod prelude;
+mod store;
+mod store_query;
 pub mod testhelpers;
 mod utils;
 
 use adapters::{CacheAdapter, CacheResult, EmitterAdapter, StoreAdapter};
-use chrono::prelude::*;
-pub use event_store_derive_internals::{EventData, Events};
+pub use aggregator::Aggregator;
+pub use event::Event;
+pub use event_context::EventContext;
+use event_store_derive_internals::Events;
 use serde::{Deserialize, Serialize};
-use serde_json::Value as JsonValue;
-use std::fmt::Debug;
+use store::Store;
+pub use store_query::StoreQuery;
 use tokio::runtime::current_thread;
-use uuid::Uuid;
-
-/// Event context
-///
-/// Contains metadata for event and, most importantly, the creation time
-#[derive(Serialize, Deserialize)]
-pub struct EventContext {
-    /// TODO: What is this?
-    action: Option<String>,
-
-    /// Optional event "subject" or metadata
-    subject: Option<JsonValue>,
-
-    /// Event creation time
-    time: DateTime<Utc>,
-}
-
-/// Event with `EventData`, `EventContext` and a `Uuid` ID
-///
-/// This is what gets stored in the store and emitted from the emitter
-#[derive(Serialize, Deserialize)]
-pub struct Event<D> {
-    data: D,
-    context: EventContext,
-    id: Uuid,
-}
-
-impl<D> Event<D>
-where
-    D: Events,
-{
-    /// Get the ID of this event
-    pub fn id(&self) -> Uuid {
-        self.id
-    }
-
-    /// Get the data of this event
-    pub fn data(&self) -> &D {
-        &self.data
-    }
-
-    /// Get the context of this event
-    pub fn context(&self) -> &EventContext {
-        &self.context
-    }
-
-    /// Create a new event
-    pub fn new(data: D, id: Uuid, context: EventContext) -> Self {
-        Self { data, context, id }
-    }
-
-    /// Create a new event from some data. `context.time` is set to now, `id` to a new V4 ID
-    ///
-    /// The rest of the context is left empty
-    pub fn from_data(data: D) -> Self {
-        Self {
-            data,
-            id: Uuid::new_v4(),
-            context: EventContext {
-                action: None,
-                subject: None,
-                time: Utc::now(),
-            },
-        }
-    }
-
-    /// Create a copied event with the given ID
-    ///
-    /// ```
-    /// # extern crate uuid;
-    /// # extern crate event_store;
-    /// # use uuid::Uuid;
-    /// # use event_store::testhelpers::*;
-    /// # use event_store::Event;
-    /// # let example_data = TestEvents::Inc(TestIncrementEvent {
-    /// #     by: 1,
-    /// #     ident: "it_aggregates_events".into(),
-    /// # });
-    /// #
-    /// let event_id = Uuid::new_v4();
-    /// let evt = Event::from_data(example_data).with_id(event_id);
-    ///
-    /// assert_eq!(evt.id(), event_id);
-    /// ```
-    pub fn with_id(self, id: Uuid) -> Self {
-        Self { id, ..self }
-    }
-}
-
-/// A query to be passed to the store
-///
-/// This trait must be implemented for whichever type you want to pass to a particular store. See
-/// impls below for examples.
-pub trait StoreQuery {}
-
-/// Aggregator trait
-///
-/// This takes three type items:
-///
-/// * `E: EventData` – The enum of domain events that items from the backing store will be parsed to
-/// * `A` – The type of the query args to use when querying the backing store. Can be as simple as
-/// a `String`, but using a `struct` is recommended for readability
-/// * `Q: StoreQuery` – The query object to pass to the backing store. It should be built from `A`
-///
-/// ### Aggregation from `E: Events` into `Self`
-///
-/// Aggregation logic is handled by `apply_event`. Every event in the resultset is passed to this
-/// function. It should return a modified copy of `Self` if a field has changed, or just `self` if
-/// the event does not change any state. For example:
-///
-/// ```rust
-/// # extern crate serde;
-/// # #[macro_use]
-/// # extern crate serde_derive;
-/// # extern crate event_store;
-/// # #[macro_use]
-/// # extern crate event_store_derive;
-/// use event_store::{StoreQuery, Event, Aggregator};
-///
-/// // An example event in this domain
-/// #[derive(Deserialize, Serialize, Debug)]
-/// struct NameChanged {
-///     name: String,
-/// }
-///
-/// // Another example event in this domain
-/// #[derive(Deserialize, Serialize, Debug)]
-/// struct EmailChanged {
-///     email: String,
-/// }
-///
-/// // We want to ignore this event; the password should never be output!
-/// #[derive(Deserialize, Serialize, Debug)]
-/// struct PasswordChanged {
-///     password: String,
-/// }
-///
-/// // Enum of all events for this domain
-/// #[derive(Events)]
-/// #[event_store(namespace = "some_namespace")]
-/// enum UsersEvents {
-///     NameChanged(NameChanged),
-///     EmailChanged(EmailChanged),
-///     PasswordChanged(PasswordChanged),
-/// }
-///
-/// // The domain entity we want to aggregate to
-/// #[derive(Clone, Debug, PartialEq)]
-/// struct UserDetails {
-///     name: String,
-///     email: String,
-/// }
-///
-/// // Initial state for user detail aggregation
-/// impl Default for UserDetails {
-///     fn default() -> Self {
-///         Self {
-///             name: "".into(),
-///             email: "".into(),
-///         }
-///     }
-/// }
-/// #
-/// # struct DummyQuery;
-/// # impl StoreQuery for DummyQuery {}
-///
-/// impl Aggregator<UsersEvents, String, DummyQuery> for UserDetails {
-///     fn apply_event(acc: Self, event: &Event<UsersEvents>) -> Self {
-///         match event.data() {
-///             UsersEvents::NameChanged(e) => Self {
-///                 name: e.name.clone(),
-///                 ..acc
-///             },
-///             UsersEvents::EmailChanged(e) => Self {
-///                 email: e.email.clone(),
-///                 ..acc
-///             },
-///             UsersEvents::PasswordChanged(_) => acc,
-///         }
-///     }
-///
-///     // ...
-/// #
-/// #     fn query(_: String) -> DummyQuery {
-/// #         DummyQuery
-/// #     }
-/// }
-///
-/// fn main() {
-/// #     let example_events = vec![
-/// #         Event::from_data(UsersEvents::NameChanged(NameChanged { name: String::from("Repositive User") })),
-/// #         Event::from_data(UsersEvents::EmailChanged(EmailChanged { email: String::from("bobby@beans.com") })),
-/// #         Event::from_data(UsersEvents::PasswordChanged(PasswordChanged { password: String::from("hunter2") })),
-/// #     ];
-/// #
-///     // Example aggregation without a store. You would normally call `store.aggregate`
-///     let user: UserDetails = example_events
-///         .iter()
-///         .fold(UserDetails::default(), UserDetails::apply_event);
-///
-///     assert_eq!(
-///         user,
-///         UserDetails { name: "Repositive User".into(), email: "bobby@beans.com".into() }
-///     );
-/// }
-/// ```
-///
-/// ### Initial state
-///
-/// `Aggregator` has a trait bound on `Default`. `Default` should be the initial state
-/// of the entity before the events are reduced onto it. For example:
-///
-/// ```rust
-/// #[derive(Clone, Debug)]
-/// struct ExampleUser {
-///     name: String,
-///     email: String,
-///     bio: Option<String>
-/// }
-///
-/// impl Default for ExampleUser {
-///     fn default() -> Self {
-///         Self {
-///             name: "".into(),
-///             email: "".into(),
-///             bio: None,
-///         }
-///     }
-/// }
-/// ```
-pub trait Aggregator<E: Events, A: Clone, Q: StoreQuery>: Clone + Debug + Default {
-    /// Apply an event `E` to `acc`, returning a copy of `Self` with updated fields. Can also just
-    /// return `acc` if nothing has changed.
-    fn apply_event(acc: Self, event: &Event<E>) -> Self;
-
-    /// Produce a query object from some query arguments
-    fn query(field: A) -> Q;
-}
-
-/// Store trait
-pub trait Store<'a, E: Events, Q: StoreQuery, S: StoreAdapter<E, Q>, C, EM: EmitterAdapter> {
-    /// Create a new event store
-    fn new(store: S, cache: C, emitter: EM) -> Self;
-
-    /// Query the backing store and return an entity `T`, reduced from queried events
-    fn aggregate<T, A>(&self, query: A) -> Result<T, String>
-    where
-        T: Aggregator<E, A, Q> + Serialize + for<'de> Deserialize<'de> + PartialEq,
-        A: Clone;
-
-    /// Save an event to the store with optional context
-    fn save(&self, event: Event<E>) -> Result<(), String>;
-}
 
 /// Main event store
 pub struct EventStore<S, C, EM> {
