+++ event-store/src/store_query.rs
@@ -1,0 +1,5 @@
+/// A query to be passed to the store
+///
+/// This trait must be implemented for whichever type you want to pass to a particular store. See
+/// impls below for examples.
+pub trait StoreQuery {}
