+++ event-store/src/testhelpers.rs
@@ -51,7 +51,7 @@ impl Default for TestCounterEntity {
 
 impl<'a> Aggregator<TestEvents, String, PgQuery<'a>> for TestCounterEntity {
     fn apply_event(acc: Self, event: &Event<TestEvents>) -> Self {
-        let counter = match event.data {
+        let counter = match event.data() {
             TestEvents::Inc(ref inc) => acc.counter + inc.by,
             TestEvents::Dec(ref dec) => acc.counter - dec.by,
         };
